requirements.txt
slowapi
aiohttp
fastapi
uvicorn[standard]
gunicorn
fastapi-jwt-auth
python-dotenv
python-decouple
sqlalchemy
sqlalchemy-utils
alembic==1.7.7
requests
email-validator
databases[postgresql]
psycopg2
bcrypt
passlib[bcrypt]